⚡️ Improve `--release` test performance
This logging is intended to be used when a test fails. Release builds
of tests may run faster, but it's expected to always use a debug
build when actually, well, debugging.

If a test starts to fail only in release builds, we have bigger
problems Those kinds of issues are generally related to UB and race
conditions, which cannot happen in pure safe code. Start debugging
in a relevant `unsafe` block.